knowhizService/requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
amqp==5.2.0
anki==24.6.3
anki-export==0.1.4
annotated-types==0.7.0
anthropic==0.31.2
anyio==4.3.0
arxiv==2.1.3
asgiref==3.8.1
attrs==23.2.0
azure-core==1.30.2
azure-storage-blob==12.21.0
backoff==2.2.1
bcrypt==4.1.3
beautifulsoup4==4.12.3
billiard==4.2.0
blinker==1.8.2
botocore==1.34.113
build==1.2.1
cachetools==5.3.3
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coloredlogs==15.0.1
cryptography==42.0.7
dataclasses-json==0.6.6
decorator==4.4.2
deepdiff==7.0.1
defusedxml==0.7.1
Deprecated==1.2.14
discord.py==2.3.2
distro==1.9.0
dnspython==2.6.1
docarray==0.32.1
email_validator==2.1.1
emoji==2.12.1
eventlet==0.36.1
faiss-cpu==1.8.0.post1
fastapi==0.99.1
fastapi-cli==0.0.4
feedparser==6.0.11
filelock==3.14.0
filetype==1.2.0
Flask==3.0.3
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.3.1
gevent==24.2.1
google-auth==2.29.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.64.0
h11==0.14.0
h2==4.1.0
hnswlib==0.8.0
hpack==4.0.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.1
humanfriendly==10.0
Hypercorn==0.16.0
hyperframe==6.0.1
idna==3.7
imageio==2.34.1
imageio-ffmpeg==0.4.9
importlib-metadata==7.0.0
importlib_resources==6.4.0
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
kombu==5.3.7
kubernetes==29.0.0
langchain==0.2.1
langchain-anthropic==0.1.20
langchain-community==0.2.1
langchain-core==0.2.22
langchain-openai==0.1.7
langchain-text-splitters==0.2.0
langdetect==1.0.9
langsmith==0.1.93
lxml==5.2.2
markdown-it-py==3.0.0
markdownify==0.13.1
MarkupSafe==2.1.5
marshmallow==3.21.2
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.16.3
openai==1.30.3
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
ordered-set==4.1.0
orjson==3.10.3
overrides==7.7.0
packaging==23.2
pandas==2.2.2
pdf2image==1.17.0
pdfminer==20191125
pdfminer.six==20231228
pillow==10.3.0
posthog==3.5.0
priority==2.0.0
proglog==0.1.10
prompt-toolkit==3.0.43
protobuf==4.25.3
pulsar-client==3.5.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==1.10.15
pydantic_core==2.18.2
pydub==0.25.1
Pygments==2.18.0
pymongo==4.7.2
PyMuPDF==1.24.4
PyMuPDFb==1.24.3
pypdf==4.2.0
PyPika==0.48.9
pyproject_hooks==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2024.4.27
python-magic==0.4.27
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
Quart==0.19.6
quart-cors==0.7.0
rapidfuzz==3.9.1
redis==5.0.4
regex==2024.5.15
requests==2.32.2
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
s3transfer==0.10.1
scikit-learn==1.5.0
scipy==1.13.1
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
sseclient-py==1.8.0
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
tenacity==8.3.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
tqdm==4.66.4
typer==0.12.3
types-requests==2.32.0.20240523
typing-inspect==0.9.0
typing_extensions==4.12.0
tzdata==2024.1
ujson==5.10.0
unstructured==0.14.2
unstructured-client==0.22.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
vine==5.1.0
watchfiles==0.21.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.3
wikipedia==1.4.0
Wikipedia-API==0.6.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
youtube-transcript-api==0.6.2
zipp==3.18.2
zope.event==5.0
zope.interface==6.4.post2
